+++ sam/models/__init__.py
@@ -1,4 +1,6 @@
-from .keras_templates import create_keras_quantile_mlp, create_keras_quantile_rnn  # noqa:F401
+from .keras_templates import create_keras_quantile_mlp, \
+    create_keras_quantile_rnn, create_keras_autoencoder_mlp, \
+    create_keras_autoencoder_rnn  # noqa: F401
 from .SamQuantileMLP import SamQuantileMLP  # noqa:F401
 from .benchmark import preprocess_data_for_benchmarking, benchmark_model  # noqa:F401
 from .benchmark import plot_score_dicts, benchmark_wrapper  # noqa:F401